Added Application Layer and Scene Example.
Modified the existing setup that was initially laid out to
now transfer the hexagon tiles to their own struct and
render everything from the HexMap app class. This will
serve as the foundation for how the application is built
going forward.

The program currently uses {x, y} coordinate scene for
identifying which hexagon is which, but this will be modified
in the future to use axial coordinates eventually. See the link
for more info: https://www.redblobgames.com/grids/hexagons/